Move remote.js to http directory.
